jira-project-config-audit/requirements.txt
flask>=2.0
mysql-connector-python>=8.0
